+++ src/bookings/bookings.controller.ts
@@ -10,7 +10,7 @@ import { MuseumIdGuard } from 'common/guards/museumId.guard';
 import { FilesValidationPipe, FileValidationPipe } from 'common/pipes/file-validation.pipe';
 import { JoiValidationPipe } from 'common/pipes/joi-validation.pipe';
 import BookingQuery from 'common/querys/booking.query';
-import { createBookingSchema, deleteBookingSchema, deleteBookingSchemaForSuperadmin, updateBookingSchema, createBookingSchemaForSuperadmin, updateBookingSchemaForSuperadmin } from 'common/schemas/booking.schema';
+import { createBookingSchema, deleteBookingSchema, deleteBookingSchemaForSuperadmin, updateBookingSchema, createBookingSchemaForSuperadmin, updateBookingSchemaForSuperadmin, payBookingSchema, generateQrCodeBookingSchema } from 'common/schemas/booking.schema';
 import { createfileGenerator, deleteFileGenerator, updatefileGenerator } from 'common/utils/image-processor.util';
 import MESSAGE from 'common/utils/message.util';
 import responseUtil from 'common/utils/response.util';
@@ -23,13 +23,19 @@ import { Cron, CronExpression } from '@nestjs/schedule';
 import dayjsUtil from 'common/utils/dayjs.util';
 import { PricesService } from 'prices/prices.service';
 import { TicketsService } from 'tickets/tickets.service';
+import { CustomException } from 'common/exceptions/custom.exception';
+import generateTransactionId from 'common/utils/transaction-generator.util';
+import generateInvoiceId from 'common/utils/inv-generator.util';
+import { Bcel } from 'bcel/bcel';
+import { PaymentsService } from 'payments/payments.service';
 
 const PREFIX = 'bookings';
 
 @Controller('')
 export class BookingsController {
     constructor(private prisma: PrismaService, private bookingsService: BookingsService,
-        private authService: AuthService, private usersService: UsersService, private pricesService: PricesService, private ticketsService: TicketsService) { }
+        private authService: AuthService, private usersService: UsersService, private pricesService: PricesService,
+        private ticketsService: TicketsService, private paymentService: PaymentsService) { }
 
     @Get("my-bookings")
     async findAll(@Req() req: Request, @Res() res: Response, @Query() { filter: { museum_id, ...filter } = {}, ...query }: BookingQuery) {
@@ -109,11 +115,136 @@ export class BookingsController {
         const jwtPayload = this.authService.jwtVerify(access_token);
         const user = await this.usersService.findById(jwtPayload["id"]);
         const data = await this.bookingsService.findById(Number(id));
+
+        if (data.museum_id !== user.museum_id) {
+            throw new CustomException({ error: MESSAGE.PermissionAccessingFailed, }, HttpStatus.FORBIDDEN);
+        }
+
         return res.json(responseUtil({
             req,
             body: user.museum_id === data.museum_id ? data : {},
         }))
     }
+
+    @UseGuards(AuthGuard)
+    @Post("bookings/qrcode-generate")
+    async generateQrCode(@Req() req: Request, @Res() res: Response,
+        @Body(new JoiValidationPipe(generateQrCodeBookingSchema)) createDto: CreateBookingDto) {
+        try {
+            const [_, access_token] = req.headers.authorization?.split(' ');
+            const jwtPayload = this.authService.jwtVerify(access_token);
+            const user = await this.usersService.findById(jwtPayload["id"]);
+            /* const dataTransactions = await this.prisma.$transaction(async (tx) => {
+                return await Promise.all(createDto.booking_details.map(async (createNestedtDto, i) => {
+                    const createFile = createfileGenerator(booking_image_paths?.[i], PREFIX, PREFIX);
+                    return {
+                        data: createNestedtDto,
+                        createFile: createFile?.generate(),
+                    }
+                }))
+            }); */
+
+            let qrCode = null;
+            let oneClickPay = null;
+            let transactionId = generateTransactionId();
+            let invoiceId = generateInvoiceId();
+            let description = `INV:${invoiceId}|USER_NAME:${user.username}|Pay`;
+            let onePay = new Bcel(transactionId, "mch62baa8923023e", "7372");
+
+            const { dto, ...createData } = await this.bookingsService.generateForQrCode({
+                ...createDto,
+                museum_id: user.museum_id,
+                user_id: user.id,
+            });
+
+            onePay.getCode({
+                transactionId,
+                invoiceId,
+                terminalId: `TIDPCK`,
+                amount: createData.total_with_discount,
+                description,
+                expireTime: 2
+            }, (code) => {
+                qrCode = `https://chart.googleapis.com/chart?chs=250x250&cht=qr&chl=${code}&choe=UTF-8`;
+                oneClickPay = `onepay://qr/${code}`;
+            })
+
+            onePay.pubnubSubscribe();
+
+            return res.status(HttpStatus.OK).json(responseUtil({
+                req, message: "Qrcode is generated", body: {
+                    booking_data: {
+                        ...dto,
+                        type: "bank",
+                        way: "booking"
+                    },
+                    bcel_details: {
+                        transaction_id: transactionId,
+                        qr_code: qrCode,
+                        one_click_pay: oneClickPay,
+                        description,
+                    }
+
+                },
+            }));
+
+        } catch (e) {
+            throw e;
+        }
+    }
+
+
+    @UseGuards(AuthGuard)
+    @Post("bookings/qrcode-pay")
+    @UseInterceptors(FileInterceptor("confirmed_image_path"))
+    async payFromQrCode(@Req() req: Request, @Res() res: Response,
+        @Body(new JoiValidationPipe(payBookingSchema)) createDto: CreateBookingDto,
+        @UploadedFile(new FileValidationPipe()) confirm_image_path: Express.Multer.File
+    ) {
+        try {
+            const [_, access_token] = req.headers.authorization?.split(' ');
+            const jwtPayload = this.authService.jwtVerify(access_token);
+            const user = await this.usersService.findById(jwtPayload["id"]);
+            const createFile = createfileGenerator(confirm_image_path, PREFIX, PREFIX);
+
+            const createData = await this.bookingsService.create({
+                ...createDto,
+                confirmed_image_path: createFile?.filePath,
+                type: "bank",
+                way: "booking",
+                status: "success",
+                paid_by_id: user.id,
+                museum_id: user.museum_id,
+            });
+
+            const payment = await this.paymentService.create({
+                booking_id: createData.id,
+                transaction_id: createDto.transaction_id,
+                invoice_id: createDto.invoice_id,
+                total: Number(createData.total),
+                bank_bill_description: createDto.description,
+                bank_percentage: 1,
+                bank_percentage_amount: Number(createData.total) * (1 / 1000)
+            })
+
+            createFile?.generate();
+
+            await Promise.all([...Array(createData.number_of_adult + createData.number_of_child).keys()].map(async () => {
+                await this.ticketsService.create({
+                    booking_id: createData.id,
+                });
+            }));
+
+            const data = await this.bookingsService.findById(createData.id);
+
+            return res.status(HttpStatus.OK).json(responseUtil({ req, message: MESSAGE.created, body: data, }));
+
+
+        } catch (e) {
+            throw e;
+        }
+    }
+
 
 
     @UseGuards(AuthGuard, MuseumIdGuard)
@@ -140,7 +271,7 @@ export class BookingsController {
                 museum_id: user.museum_id,
                 user_id: user.id,
             });
-            
+
             await Promise.all([...Array(createData.number_of_adult + createData.number_of_child).keys()].map(async () => {
                 await this.ticketsService.create({
                     booking_id: createData.id,
@@ -162,6 +293,15 @@ export class BookingsController {
     async update(@Req() req: Request, @Res() res: Response,
         @Body(new JoiValidationPipe(updateBookingSchema)) updateDto: UpdateBookingDto,) {
         try {
+            const [_, access_token] = req.headers.authorization?.split(' ');
+            const jwtPayload = await this.authService.jwtDecode(access_token);
+            const user = await this.usersService.findById(jwtPayload["id"]);
+            const dataById = await this.bookingsService.findById(updateDto.id);
+
+            if (dataById.museum_id !== user.museum_id) {
+                throw new CustomException({ error: MESSAGE.PermissionAccessingFailed, }, HttpStatus.FORBIDDEN);
+            }
+
             const data = await this.bookingsService.update({
                 ...updateDto,
             });
@@ -175,30 +315,19 @@ export class BookingsController {
     }
 
     @UseGuards(AuthGuard, MuseumIdGuard)
-    @Roles(Role.ADMIN, Role.MANAGER, Role.GOD, Role.OWNER)
-    @Delete("admin/bookings/:id")
-    async delete(@Req() req: Request, @Res() res: Response,
-        @Param('id') id: string) {
-        try {
-            const dataById = await this.bookingsService.findById(parseInt(id));
-            const data = await this.bookingsService.delete({
-                id: dataById.id,
-            });
-
-            return res.status(HttpStatus.OK).json(responseUtil({ req, message: MESSAGE.deleted, body: data }));
-
-        } catch (e) {
-            throw e;
-        }
-    }
-
-    @UseGuards(AuthGuard, MuseumIdGuard)
     @Roles(Role.ADMIN, Role.MANAGER, Role.GOD, Role.OWNER, Role.CASHIER)
     @Delete("admin/bookings/:id")
     async deleteForAdmin(@Req() req: Request, @Res() res: Response,
         @Param('id') id: string) {
         try {
+            const [_, access_token] = req.headers.authorization?.split(' ');
+            const jwtPayload = await this.authService.jwtDecode(access_token);
+            const user = await this.usersService.findById(jwtPayload["id"]);
             const dataById = await this.bookingsService.findById(parseInt(id));
+
+            if (dataById.museum_id !== user.museum_id) {
+                throw new CustomException({ error: MESSAGE.PermissionAccessingFailed, }, HttpStatus.FORBIDDEN);
+            }
             const data = await this.bookingsService.delete({
                 id: dataById.id,
             });
@@ -249,7 +378,7 @@ export class BookingsController {
             const createData = await this.bookingsService.create({
                 ...createDto,
             });
-            
+
             await Promise.all([...Array(createData.number_of_adult + createData.number_of_child).keys()].map(async () => {
                 await this.ticketsService.create({
                     booking_id: createData.id,
